+++ CHANGELOG.md
@@ -8,6 +8,7 @@ Other Changes
 
 - Added colorization of items in the status bar on hover to better fit with the color that Peacock applies.
 - Added colorization of the Activity Bar badge based on the background color applied by Peacock to improve readability
+- Modified inactive element coloring to better match default behavior of VS Code and most themes.
 
 ## 0.7.0